functions/gdelt_sync/requirements.txt
google-cloud-bigquery==3.14.1
google-cloud-secret-manager==2.16.4
google-cloud-pubsub==2.18.4
functions-framework==3.5.0
Flask==3.0.0
pytz==2023.3
